+++ node/libs/storage/src/traits.rs
@@ -5,38 +5,6 @@ use std::{sync::Arc, fmt, ops};
 use zksync_concurrency::{ctx};
 use zksync_consensus_roles::validator::{BlockNumber, FinalBlock, Payload};
 
-/// Storage of a continuous range of L2 blocks.
-///
-/// Implementations **must** propagate context cancellation using [`StorageError::Canceled`].
-#[async_trait]
-pub trait BlockStore: fmt::Debug + Send + Sync {
-    /// Range of blocks avaliable in storage.
-    /// Consensus code calls this method only once and then tracks the
-    /// range of avaliable blocks internally.
-    async fn available_blocks(&self, ctx: &ctx::Ctx) -> ctx::Result<ops::Range<BlockNumber>>;
-
-    /// Gets a block by its number. Should returns an error if block is not available.
-    async fn block(&self, ctx: &ctx::Ctx, number: BlockNumber) -> ctx::Result<FinalBlock>;
-
-    /// Persistently store a block.
-    /// Implementations are only required to accept a block directly after the current last block,
-    /// so that the stored blocks always constitute a continuous range.
-    /// Implementation should return only after the block is stored PERSISTENTLY -
-    /// consensus liveness property depends on this behavior.
-    async fn store_next_block(&self, ctx: &ctx::Ctx, block: &FinalBlock) -> ctx::Result<()>;
-}
-
-/// Storage for [`ReplicaState`].
-///
-/// Implementations **must** propagate context cancellation using [`StorageError::Canceled`].
-#[async_trait]
-pub trait ReplicaStore: fmt::Debug + Send + Sync {
-    /// Gets the replica state, if it is contained in the database.
-    async fn state(&self, ctx: &ctx::Ctx) -> ctx::Result<Option<ReplicaState>>;
-
-    /// Stores the given replica state into the database.
-    async fn set_state(&self, ctx: &ctx::Ctx, replica_state: &ReplicaState) -> ctx::Result<()>;
-}
 
 #[async_trait]
 pub trait ValidatorStore {
